[code_generation] Those changes should help for now.
Because currently having input models in query elements is broken, I'll use the `Json` type, which is broken as well when used like `Json[Model]`.
The only configuration to make it work is to use plain `Json` without any type. Thus for now doing exactly that, but the code doesn't need to know, so there's a class overwriting `Json` and not yet providing the `[Model]` part to the pydantic `Json`.

That should work until https://github.com/tiangolo/fastapi/issues/884 is resolved.